chore: Move code to dedicated helpers
